update UI for workspace settings pfp
